% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_participation.R
\name{explore_participation}
\alias{explore_participation}
\title{Detailed Panel Data Participation Patterns}
\usage{
explore_participation(data, group, time, max_patterns = 10)
}
\arguments{
\item{data}{A data.frame containing panel data.}

\item{group}{A character string specifying the name of the entity/group variable in panel data.}

\item{time}{A character string specifying the name of the time variable.}

\item{max_patterns}{An integer specifying the maximum number of patterns to display in detail. Default = 10.}
}
\value{
A list containing detailed participation pattern statistics.
The list includes:
\item{pattern_matrix}{Matrix showing presence patterns}
\item{pattern_groups}{List of entities belonging to each pattern}
\item{pattern_stats}{Data.frame with pattern statistics}
\item{entities_by_pattern}{Vector with the number of entities in each pattern}
The result is returned invisibly, and will be printed automatically unless assigned to a variable.
}
\description{
Provides detailed analysis of participation patterns in panel data.
}
\examples{
data(production)

# Basic usage - will print automatically
explore_participation(production, group = "firm", time = "year")

# Show only top 5 patterns - will print automatically
explore_participation(production, group = "firm", time = "year", max_patterns = 5)

# Access the number of entities per pattern - won't print
result <- explore_participation(production, group = "firm", time = "year")
entities_per_pattern <- result$entities_by_pattern
print(entities_per_pattern)

}
