% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_panel.R
\name{describe_panel}
\alias{describe_panel}
\title{Panel Data Structure Description}
\usage{
describe_panel(data, group, time)
}
\arguments{
\item{data}{A data.frame containing panel data.}

\item{group}{A character string specifying the name of the entity/group variable in panel data.}

\item{time}{A character string specifying the name of the time variable in panel data.}
}
\value{
A data.frame with three columns:
\describe{
\item{\code{entities}}{Number of unique groups/entities in the panel}
\item{\code{time_periods}}{Number of unique time periods in the panel}
\item{\code{observations}}{Total number of observations in the panel}
}
}
\description{
This function provides basic summary information about panel data structure.
}
\details{
This function provides basic panel data structure information including
the number of unique entities, time periods, and total observations.
}
\examples{
data(production)

# Get panel description
panel_desc <- describe_panel(production, group = "firm", time = "year")
print(panel_desc)

}
\seealso{
\code{\link[=explore_panel]{explore_panel()}} for detailed panel analysis
}
