% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_panel.R
\name{describe_panel}
\alias{describe_panel}
\title{Panel Data Structure Description}
\usage{
describe_panel(data, group = NULL, time = NULL)
}
\arguments{
\item{data}{A data.frame containing panel data, or a data.frame with panel attributes.}

\item{group}{A character string specifying the name of the entity/group variable
in panel data. Not required if data has panel attributes.}

\item{time}{A character string specifying the name of the time variable in
panel data. Not required if data has panel attributes.}
}
\value{
A data.frame with three columns:
\describe{
\item{\code{observations}}{Total number of observations in the panel}
\item{\code{entities}}{Number of unique groups/entities in the panel}
\item{\code{periods}}{Number of unique time periods in the panel}
}

The data.frame has additional attributes:
\describe{
\item{\code{panel_group}}{The grouping variable name}
\item{\code{panel_time}}{The time variable name}
}
}
\description{
This function provides basic summary information about panel data structure.
}
\details{
This function provides basic panel data structure information including
the number of unique entities, time periods, and total observations.
When provided with a data.frame that has panel attributes (created by set_panel()),
the function automatically extracts group and time variable names from the attributes.
}
\examples{
data(production)

# Method 1: With regular data.frame
panel_desc <- describe_panel(production, group = "firm", time = "year")
print(panel_desc)

# Method 2: With data.frame with panel attributes
panel_data <- set_panel(production, group = "firm", time = "year")
panel_desc <- describe_panel(panel_data)
print(panel_desc)

}
\seealso{
\code{\link[=explore_panel]{explore_panel()}}, \code{\link[=set_panel]{set_panel()}}
}
