% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_periods.R
\name{describe_periods}
\alias{describe_periods}
\title{Time Period Coverage Description}
\usage{
describe_periods(data, group = NULL, time = NULL)
}
\arguments{
\item{data}{A data.frame containing panel data, or a data.frame with panel attributes.}

\item{group}{A character string specifying the name of the entity/group variable in panel data.
Not required if data has panel attributes.}

\item{time}{A character string specifying the name of the time variable.
Not required if data has panel attributes.}
}
\value{
A data.frame with time period coverage statistics.
}
\description{
Provides detailed coverage statistics for each time period in panel data.
}
\details{
The data.frame contains the following columns:
\describe{
\item{\code{[time]}}{Time period identifier (name matches the input \code{time} argument)}
\item{\code{total}}{Total number of observations in each period (all rows)}
\item{\code{balanced}}{Number of observations with at least one non-NA value in substantive variables}
\item{\code{complete}}{Number of observations without any NA values in substantive variables}
}

Time periods are sorted naturally (numeric values as numbers, others alphabetically).
}
\examples{
data(production)
describe_periods(production, group = "firm", time = "year")

# With panel attributes
panel_data <- set_panel(production, group = "firm", time = "year")
describe_periods(panel_data)

}
\seealso{
\code{\link[=describe_balance]{describe_balance()}}, \code{\link[=explore_balance]{explore_balance()}}, \code{\link[=describe_participation]{describe_participation()}}
}
