% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_panel.R
\name{set_panel}
\alias{set_panel}
\title{Add Panel Attributes to Data}
\usage{
set_panel(data, group, time)
}
\arguments{
\item{data}{A data.frame containing panel data.}

\item{group}{A character string specifying the name of the entity/group variable.}

\item{time}{A character string specifying the name of the time variable.}
}
\value{
The input data.frame with two additional attributes: panel_group and panel_time.
}
\description{
Adds panel structure attributes to a data.frame, storing group and time variable names.
This allows panel functions to automatically detect panel structure.
}
\details{
This function adds attributes to a data.frame to mark it as panel data.
The attributes store the names of group and time variables, enabling
panel-aware functions to automatically detect the panel structure without
requiring explicit specification of group and time variables in each function call.
The original data.frame structure is preserved.
}
\examples{
data(production)

# Add panel attributes
panel_data <- set_panel(production, group = "firm", time = "year")

# Check the attributes
attr(panel_data, "panel_group")
attr(panel_data, "panel_time")

# Use with describe_panel()
describe_panel(panel_data)

}
\seealso{
\code{\link[=check_panel]{check_panel()}}, \code{\link[=get_attributes]{get_attributes()}}, \code{\link[=describe_panel]{describe_panel()}}, \code{\link[=explore_panel]{explore_panel()}}
}
